requirements.txt
Django>=4.2,<5.0
google-generativeai>=0.5.0
pypdf>=4.0.0
python-dotenv>=1.0.0
